Started Day 5
Presently, my approach to part 1 is too slow.
I might have to compress the information contained in the almanac.
To do so, I will need to know the bounds of the numbers.
I assume the least is 1, but the greatest could be arbitrarily large.
Also, I should probably not construct so many lists out of ranges.
Instead, apply arithmetic.
I should determine how to compose mappings
without separately invoking the component functions.